+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/ConfirmDelete.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "@i18n/client";
 import { ConfirmFormDeleteDialog } from "./shared";
 import { toast, ToastContainer } from "./shared/Toast";
 import axios, { AxiosError, AxiosResponse } from "axios";
-import { clearTemplateStore } from "@lib/store/useTemplateStore";
+import { clearTemplateStore } from "@lib/store/utils";
 
 const handleDelete = async (formID?: string): Promise<AxiosResponse | { error: AxiosError }> => {
   try {
