+++ app/(gcforms)/[locale]/(user authentication)/components/client/LoginMenu.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "@i18n/client";
 import Link from "next/link";
 import { logout } from "../action";
 import { useSession } from "next-auth/react";
-import { clearTemplateStore } from "@lib/store/useTemplateStore";
+import { clearTemplateStore } from "@lib/store/utils";
 
 export const LoginMenu = () => {
   const { i18n, t } = useTranslation("common");
