+++ components/clientComponents/globals/Header/YourAccountDropdown.tsx
@@ -5,7 +5,7 @@ import { ChevronDown } from "@serverComponents/icons";
 import { useTranslation } from "@i18n/client";
 import Link from "next/link";
 import { signOut } from "next-auth/react";
-import { clearTemplateStore } from "@lib/store/useTemplateStore";
+import { clearTemplateStore } from "@lib/store/utils";
 import { useAccessControl } from "@lib/hooks/useAccessControl";
 
 type YourAccountDropdownProps = {
